@@ -12,42 +12,42 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.2.7",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^8.2.7",
-    "@angular/compiler": "^5.2.11",
-    "@angular/core": "^8.2.7",
-    "@angular/forms": "^8.2.7",
+    "@angular/animations": "~8.2.7",
+    "@angular/cdk": "^8.2.0",
+    "@angular/common": "~8.2.7",
+    "@angular/compiler": "~8.2.7",
+    "@angular/core": "~8.2.7",
+    "@angular/forms": "~8.2.7",
     "@angular/http": "^7.2.15",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^8.2.7",
-    "@angular/platform-browser-dynamic": "^8.2.7",
-    "@angular/router": "^8.2.7",
-    "core-js": "^2.6.9",
-    "materialize-css": "^1.0.0-rc.2",
+    "@angular/material": "^8.2.0",
+    "@angular/platform-browser": "~8.2.7",
+    "@angular/platform-browser-dynamic": "~8.2.7",
+    "@angular/router": "~8.2.7",
+    "core-js": "^3.2.1",
     "ng2-cookies": "^1.0.12",
     "rxjs": "^6.5.3",
-    "zone.js": "^0.10.2"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "^8.3.5",
-    "@angular/compiler-cli": "^8.2.7",
-    "@angular/language-service": "^5.2.11",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "^2.0.6",
-    "@types/node": "^6.14.7",
-    "codelyzer": "^5.1.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "^4.3.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.4.3",
-    "karma-jasmine": "^1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "^6.0.0",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "^2.4.2"
+    "@angular-devkit/build-angular": "~0.803.5",
+    "@angular/cli": "~8.3.5",
+    "@angular/compiler-cli": "~8.2.7",
+    "@angular/language-service": "~8.2.7",
+    "@types/node": "~8.9.4",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.4.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.1.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.5.3"
   }
 }
