@@ -1,8 +1,8 @@
 ﻿import { Injectable } from '@angular/core';
 import { Router, NavigationStart } from '@angular/router';
 // tslint:disable-next-line:import-blacklist
-import { Observable } from 'rxjs';
-import { Subject } from 'rxjs/Subject';
+import {Observable, Subject} from 'rxjs';
+
 
 @Injectable()
 export class AlertService {
