@@ -1,11 +1,8 @@
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
 import { Cookie } from 'ng2-cookies';
-import { Http, Response, Headers, RequestOptions } from '@angular/http';
-import { Observable } from 'rxjs';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/map';
-import { Usuario } from '../_models/net/mrsistemas/index';
+import {HttpHeaders, HttpClient} from '@angular/common/http';
+import { environment } from '../../environments/environment';
 
 export class Foo {
   constructor(
@@ -15,28 +12,52 @@ export class Foo {
 
 @Injectable()
 export class LoginService {
+  private env = environment;
 
   constructor(
-    private _router: Router, private _http: Http) { }
+    private _router: Router, private _http: HttpClient) { }
 
   obtainAccessToken(loginData: { username: any; password: any; }) {
     const params = new URLSearchParams();
     params.append('username', loginData.username);
     params.append('password', loginData.password);
     params.append('grant_type', 'password');
-    params.append('client_id', 'USER_CLIENT_APP');
+    params.append('client_id', this.env.appId);
 
-    const headers = new Headers({ 'Content-type': 'application/x-www-form-urlencoded; charset=utf-8',
-    'Authorization': 'Basic ' + btoa('USER_CLIENT_APP:password') });
-    const options = new RequestOptions({ headers: headers });
-    this._http.post('http://localhost:8080/healthyOauthServer/oauth/token', params.toString(), options)
-      .map(res => res.json())
+    const httpOptions = {
+      headers: new HttpHeaders({
+        'Content-type': 'application/x-www-form-urlencoded; charset=utf-8',
+        'Authorization': 'Basic ' + btoa(this.env.appConnect)
+      })
+    };
+
+    this._http.post(this.env.authenticatedUri + this.env.token, params.toString(), httpOptions)
       .subscribe(
-        data => {
-          console.log('Data', data);
+        (data : any) => {
+          console.log('DataAuthenticated', data);
+          console.log('name: ', data.id)
           this.saveToken(data);
+
+          const httpAuthenticated = {
+            headers: new HttpHeaders({
+              'Content-type': 'application/x-www-form-urlencoded; charset=utf-8',
+              'Authorization': 'Bearer ' + Cookie.get('access_token')
+            })
+          }
+
+          this._http.get(this.env.clientDetailsURL + this.env.userUri + this.env.userByIdMethod + data.id, httpAuthenticated)
+            .subscribe(
+              (data : any) => {
+                console.log('DataUsers', data);
+              },
+              error => {
+                console.log('Hay un error' ,error)
+              })
         },
-        err => alert('Invalid Credentials')
+        error => {
+          alert('Invalid Credentials');
+          console.log('Hay un error' ,error)
+        }
       );
   }
 
@@ -46,13 +67,30 @@ export class LoginService {
     this._router.navigate(['/dashboard']);
   }
 
-  getResource(resourceUrl: string): Observable<Foo> {
+  saveData(token: { expires_in: number; access_token: string; }) {
+    const expireDate = new Date().getTime() + (1000 * token.expires_in);
+    Cookie.set('access_token', token.access_token, expireDate);
+    this._router.navigate(['/dashboard']);
+  }
+
+  getResource(resourceUrl: string) {
     // tslint:disable-next-line:max-line-length
-    const headers = new Headers({ 'Content-type': 'application/x-www-form-urlencoded; charset=utf-8', 'Authorization': 'Bearer ' + Cookie.get('access_token') });
-    const options = new RequestOptions({ headers: headers });
-    return this._http.get(resourceUrl, options)
-      .map((res: Response) => res.json())
-      .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+    //const headers = new Headers({ 'Content-type': 'application/x-www-form-urlencoded; charset=utf-8', 'Authorization': 'Bearer ' + Cookie.get('access_token') });
+    const httpOptions = {
+      headers: new HttpHeaders({
+        'Content-type': 'application/x-www-form-urlencoded; charset=utf-8',
+        'Authorization': 'Bearer ' + Cookie.get('access_token')
+      })
+    };
+    return this._http.get(resourceUrl, httpOptions)
+      .subscribe(
+        (res: Response) => {
+          console.log(res.json());
+        },
+      error => {
+          console.log('Error', error)
+      }
+      )
   }
 
   checkCredentials() {
