@@ -1,6 +1,6 @@
 ﻿import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { Http, Response, Headers, RequestOptions } from '@angular/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
+
 import { Cookie } from 'ng2-cookies';
 
 import { User, Usuario, Credentials } from '../_models/net/mrsistemas/index';
@@ -14,42 +14,51 @@ import { environment } from '../../environments/environment';
 @Injectable()
 export class UserService {
     public  credentials: User;
-    constructor(private _http: Http,
+    constructor(private _http: HttpClient,
         private http: HttpClient) { }
 
     getAll() {
-        console.log('Consultando datos....');
-        const headers = new Headers({ 'Content-type': 'application/x-www-form-urlencoded; charset=utf-8',
-        'Authorization': 'Bearer ' + Cookie.get('access_token') });
-        const options = new RequestOptions({ headers: headers });
-        return this._http.get(environment.userUri + environment.extraMethod , options)
-          .map((res: Response) => {
-
-              console.log('Respuesta data' + res.json());
-            })
-          .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
-    }
-
-    getResource(): Observable<Credentials> {
-        const headers = new Headers({ 'Content-type': 'application/x-www-form-urlencoded; charset=utf-8',
-        'Authorization': 'Bearer ' + Cookie.get('access_token') });
-
-        const options = new RequestOptions({ headers: headers });
-        return this._http.get(environment.userUri + environment.extraMethod, options)
-        .map((res: Response) => res.json())
-        .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+        const httpOptions = {
+        headers: new HttpHeaders({
+          'Content-type': 'application/x-www-form-urlencoded; charset=utf-8',
+          'Authorization': 'Bearer ' + Cookie.get('access_token')
+        })
       }
 
-    getById(id: number): Observable<User> {
-        const headers = new Headers({ 'Content-type': 'application/x-www-form-urlencoded; charset=utf-8',
-        'Authorization': 'Bearer ' + Cookie.get('access_token') });
+        return this._http.get(environment.userUri + environment.extraMethod , httpOptions)
+          .subscribe((res: any) => {
 
-        const options = new RequestOptions({ headers: headers });
-        return this._http.get(environment.userUri + environment.userByIdMethod + id, options)
-        .map((res: Response) => {res.json();
+              console.log('Respuesta data' + res.json());
+            },
+            (error: any) => {error.json().error || 'Server error'});
+    }
+
+    getResource() {
+      const httpOptions = {
+        headers: new HttpHeaders({
+          'Content-type': 'application/x-www-form-urlencoded; charset=utf-8',
+          'Authorization': 'Bearer ' + Cookie.get('access_token')
+        })
+      }
+
+        return this._http.get(environment.userUri + environment.extraMethod, httpOptions)
+        .subscribe((res: Response) => {
+          return res.json()
+          }, (error: any) => {console.log('error' , error)});
+      }
+
+    getById(id: number) {
+      const httpOptions = {
+        headers: new HttpHeaders({
+          'Content-type': 'application/x-www-form-urlencoded; charset=utf-8',
+          'Authorization': 'Bearer ' + Cookie.get('access_token')
+        })
+      }
+        return this._http.get(environment.userUri + environment.userByIdMethod + id, httpOptions)
+        .subscribe((res: any) => {
             this.credentials = res.json();
-        return res.json(); })
-        .catch((error: any) => Observable.throw(error.json().error || 'Server error'));
+        },
+          (error: any) => Observable.throw(error.json().error || 'Server error'));
     }
 
     create(user: User) {
