@@ -4,15 +4,12 @@ import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule } from '@angular/forms';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 
-// used to create fake backend
-import { fakeBackendProvider } from './_helpers/index';
-
 import { AppComponent } from './app.component';
 import { routing } from './app.routing';
 
 import { AlertComponent } from './_directives/index';
 import { AuthGuard } from './_guards/index';
-import { JwtInterceptor } from './_helpers/index';
+
 import { AlertService, AuthenticationService, UserService } from './_services/index';
 import { HomeComponent } from './home/index';
 import { LoginComponent } from './component/login/index';
@@ -20,14 +17,12 @@ import { RegisterComponent } from './component/register/index';
 import { HeaderComponent } from './component/header/index';
 import { FooterComponent } from './component/footer/index';
 import { DashboardComponent } from './component/dashboard/index';
-import { HttpModule } from '@angular/http';
-
+import {JwtInterceptor} from "./_helpers";
 
 @NgModule({
     imports: [
         BrowserModule,
         FormsModule,
-        HttpModule,
         HttpClientModule,
         routing
     ],
@@ -51,9 +46,7 @@ import { HttpModule } from '@angular/http';
             provide: HTTP_INTERCEPTORS,
             useClass: JwtInterceptor,
             multi: true
-        },
-        // provider used to create fake backend
-        fakeBackendProvider
+        }
     ],
     bootstrap: [AppComponent]
 })
